do not use PRIuPTR, doesn't work on Mac
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2783 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/ArticleView/ArticleTextViewPlugin.m
@@ -133,7 +133,7 @@
          atIndex: (NSUInteger) charIndex
 {
     BOOL result = NO;
-    NSLog(@"textView:clickedOnLink: %@ atIndex: %"PRIuPTR"", link, charIndex);
+    NSLog(@"textView:clickedOnLink: %@ atIndex: %lu", link, (unsigned long)charIndex);
     
     if ([link isKindOfClass: [NSURL class]]) {
         result = [[NSWorkspace sharedWorkspace] openURL: (NSURL*)link];
